+++ benchmarks/Transformers/BM25TransformerBench.php
@@ -41,7 +41,7 @@ class BM25TransformerBench
     /**
      * @Subject
      * @Iterations(3)
-     * @OutputTimeUnit("seconds", precision=3)
+     * @OutputTimeUnit("milliseconds", precision=3)
      */
     public function apply() : void
     {